[wip] Split `secp256k1` module into submodules
* Add a PoC canonical signing
